requirements.txt
aiogram==3.10.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.141.0
openai==1.40.2
pillow==10.4.0
httpx==0.27.0
alembic
pandas
openpyxl